LayoutTests/imported/w3c/web-platform-tests/cookie-store/serviceworker_cookieStore_cross_origin.https.sub-expected.txt


FAIL cookieStore.get() in ServiceWorker reads cookie set in cross-origin frame assert_equals: expected 1 but got 0
